requirements.txt
m3u8
pycryptodome
requests